// src/client/main/infra/apis/api-core/models/FetchExternalAtaDetailResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type FetchExternalAtaDetailResponse = {
  numeroAtaRegistroPreco?: string;
  anoAta?: number;
  sequencialAta?: number;
  numeroControlePNCP?: string;
  numeroControlePncpCompra?: string;
  objetoCompra?: string;
  informacaoComplementarCompra?: (string | null);
  modalidadeNome?: string;
  dataAssinatura?: string;
  dataVigenciaInicio?: string;
  dataVigenciaFim?: string;
  dataCancelamento?: (string | null);
  cancelado?: boolean;
  dataPublicacaoPncp?: string;
  dataInclusao?: string;
  dataAtualizacao?: string;
  orgaoCnpj?: string;
  orgaoNome?: string;
  unidadeNome?: string;
  unidadeCodigo?: string;
  municipioNome?: string;
  ufSigla?: string;
};