// src/client/main/infra/apis/api-core/models/FetchExternalAtaFilesResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type FetchExternalAtaFilesResponse = {
  files: Array<{
    url?: string;
    titulo?: string;
    tipoDocumentoNome?: string;
    sequencialDocumento?: number;
    dataPublicacaoPncp?: string;
  }>;
};