// src/client/main/infra/apis/api-core/models/FetchExternalContractFilesResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type FetchExternalContractFilesResponse = {
  files: Array<{
    url?: string;
    titulo?: string;
    tipoDocumentoNome?: string;
    sequencialDocumento?: number;
    dataPublicacaoPncp?: string;
  }>;
};