// src/client/main/infra/apis/api-core/models/FetchExternalProcurementContractsResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type FetchExternalProcurementContractsResponse = {
  contracts: Array<{
    numeroContratoEmpenho?: string;
    anoContrato?: number;
    sequencialContrato?: number;
    numeroControlePNCP?: string;
    objetoContrato?: string;
    niFornecedor?: string;
    nomeRazaoSocialFornecedor?: string;
    tipoPessoa?: string;
    dataAssinatura?: string;
    dataVigenciaInicio?: string;
    dataVigenciaFim?: string;
    dataPublicacaoPncp?: string;
    valorInicial?: (number | null);
    valorGlobal?: (number | null);
    valorAcumulado?: (number | null);
    tipoContratoNome?: string;
    informacaoComplementar?: (string | null);
  }>;
};