// src/client/main/infra/apis/api-core/models/FetchExternalProcurementFilesResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type FetchExternalProcurementFilesResponse = {
  files: Array<{
    url?: string;
    titulo?: string;
    tipoDocumentoNome?: string;
    sequencialDocumento?: number;
    dataPublicacaoPncp?: string;
  }>;
};