// src/client/main/infra/apis/api-core/models/FetchExternalProcurementItemsResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type FetchExternalProcurementItemsResponse = {
  items: Array<{
    numeroItem?: number;
    descricao?: string;
    quantidade?: (number | null);
    unidadeMedida?: string;
    valorUnitarioEstimado?: (number | null);
    valorTotal?: (number | null);
    materialOuServico?: string;
    materialOuServicoNome?: string;
    situacaoCompraItemNome?: string;
    criterioJulgamentoNome?: string;
    tipoBeneficioNome?: string;
    orcamentoSigiloso?: boolean;
    informacaoComplementar?: (string | null);
  }>;
  total: number;
};